+++ org.mixedrealitytoolkit.windowsspeech/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## [Unreleased]
+
+### Changed
+
+* Made the dependency on com.microsoft.mrtk.tts.windows optional. [PR #1060](https://github.com/MixedRealityToolkit/MixedRealityToolkit-Unity/pull/1060)
+
 ## [3.0.4] - 2025-11-12
 
 ### Changed